# api/openapi.yaml
openapi: 3.1.0
info:
  title: RockServer API
  version: 0.5.0
  description: >-
    Contract for RockServer operational health checks, deterministic station search, and the
    Windows voice-command JSON and streaming boundaries. `POST /api/v1/voice/command` resolves
    an already-recognized command and `GET /api/v1/voice/stream` upgrades to the
    provider-neutral WebSocket protocol for raw PCM audio and incremental transcripts.
    PostgreSQL may use provenance-compatible semantic ranking; metadata fallback preserves
    the same schemas when embeddings are absent or unavailable. Operations marked
    `x-rockserver-status: planned` are contract-first device-control designs and are not yet
    registered by the runtime. The server-resolved `station.play_stream` dispatch, device voice
    sessions, and the two device-facing
    catalog paths (`GET /api/v1/device-control/catalog/stations` and
    `GET /api/v1/device-control/catalog/search`) are implemented behind the RockserverBearer
    device session.
servers:
  - url: https://alex.vault57.ru
  - url: http://localhost
paths:
...
  /api/v1/admin/icons/import:
    get:
      operationId: getLatestStationIconImport
      summary: Read the latest durable station-icon import progress for the administrator console
      security: [{ AdminBearer: [] }]
      responses:
        "200":
          description: Latest safe counters, or null when no import has been started.
          content:
            application/json: { schema: { oneOf: [{ $ref: "#/components/schemas/StationIconImportProgress" }, { type: "null" }] } }
        "401": { description: Bearer session is absent, expired, revoked, malformed, or invalid. }
        "503": { description: Persistent icon storage or import metadata is unavailable. }
    post:
      operationId: startStationIconImport
      summary: Start one administrator-authorized station-icon import snapshot
      description: >-
        Only a trusted first-party administrator browser may start this durable background job.
        The response contains safe persisted counters; source URLs and image bytes are never returned.
      security: [{ AdminBearer: [] }]
      responses:
        "202":
          description: The import snapshot was created and its background worker was started.
          content:
            application/json: { schema: { $ref: "#/components/schemas/StationIconImportProgress" } }
        "401": { description: Bearer session is absent, expired, revoked, malformed, or invalid. }
        "403": { description: Request origin is not the trusted first-party administrator application. }
        "503": { description: Persistent icon storage or import metadata is unavailable. }
  /api/v1/admin/icons/import/{job_id}:
    get:
      operationId: getStationIconImportProgress
      summary: Read safe persisted counters for one station-icon import job
      security: [{ AdminBearer: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current or terminal progress counters.
          content:
            application/json: { schema: { $ref: "#/components/schemas/StationIconImportProgress" } }
        "401": { description: Bearer session is absent, expired, revoked, malformed, or invalid. }
        "404": { description: The job does not exist. }
        "503": { description: Persistent icon storage or import metadata is unavailable. }
  /api/v1/admin/stations/{station_id}/icon:
    put:
      operationId: replaceManualStationIcon
      summary: Upload or replace one manual station-icon override
      description: >-
        The trusted first-party administrator browser sends one raw PNG, JPEG, WebP or ICO body.
        RockServer applies the same two-MiB/signature/decode limits as automatic import and stores
        only the prepared WebP result. The automatic source URL is retained but cannot overwrite
        this manual override.
      security: [{ AdminBearer: [] }]
      parameters:
        - name: station_id
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      requestBody:
        required: true
        content:
          image/png: { schema: { type: string, format: binary } }
          image/jpeg: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
          image/x-icon: { schema: { type: string, format: binary } }
      responses:
        "204": { description: The prepared manual override was atomically published. }
        "400": { description: Station identifier is invalid. }
        "401": { description: Bearer session is absent, expired, revoked, malformed, or invalid. }
        "403": { description: Request origin is not the trusted first-party administrator application. }
        "404": { description: Station was not found. }
        "413": { description: The raw upload exceeds two MiB. }
        "422": { description: Image signature, decoding or dimensions are invalid. }
        "503": { description: Persistent icon storage or metadata is unavailable. }
    delete:
      operationId: removeManualStationIcon
      summary: Remove one manual station-icon override after explicit confirmation
      security: [{ AdminBearer: [] }]
      parameters:
        - name: station_id
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmation]
              properties:
                confirmation: { type: string, const: DELETE }
      responses:
        "204": { description: Manual override was detached; a future explicit import may use the retained automatic source. }
        "400": { description: Station identifier is invalid. }
        "401": { description: Bearer session is absent, expired, revoked, malformed, or invalid. }
        "403": { description: Request origin is not the trusted first-party administrator application. }
        "404": { description: No manual override exists for the station. }
        "422": { description: Confirmation is missing or invalid. }
        "503": { description: Persistent icon storage or metadata is unavailable. }
...
components:
...
  securitySchemes:
    RockserverBrowserCookie:
      type: apiKey
      in: cookie
      name: rockserver_browser
      description: >-
        Secure, HttpOnly, SameSite browser session cookie issued after a verified passkey ceremony.
        State-changing browser requests also require the X-CSRF-Token header and the trusted Caddy
        proxy proof; those transport checks are enforced outside the bearer scheme.
    RockCastBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Legacy deployment-configured Bearer accepted when supplied to `/api/v1/voice/*` routes.
        Normal `/api/v1/search` and `/api/v1/voice/*` product routes are anonymous; this credential is not
        part of the account/device pairing UX and must not be exposed there.
    RockserverBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Short-lived opaque native-session access token issued by the existing pairing/session flow.
        It is the sole identity accepted by planned device-control operations and is also accepted
        by current native account and voice routes. Roles/capabilities are not token scopes;
        device-control policy is computed and enforced server-side.
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived, server-revocable administrator session token. It is separate from native and browser account credentials.
...
  schemas:
...
    StationIconImportProgress:
      type: object
      additionalProperties: false
      required: [id, status, selected, processed, ready, missing, retryable_error, permanent_error, skipped]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [running, completed, interrupted, failed] }
        selected: { type: integer, minimum: 0 }
        processed: { type: integer, minimum: 0 }
        ready: { type: integer, minimum: 0 }
        missing: { type: integer, minimum: 0 }
        retryable_error: { type: integer, minimum: 0 }
        permanent_error: { type: integer, minimum: 0 }
        skipped: { type: integer, minimum: 0 }
...